Parse CSV in raw mode and benchmark
We first used the following code to benchmark stdout performance:

    let out = io::stdout();
    let mut locked_out = out.lock();

    let chunks = 1000000000;
    let data = "The quick brown fox jumped over the lazy dog\n";
    for _ in 0..chunks {
        locked_out.write_all(data.as_bytes()).unwrap()
    }

To benchmark, we piped `cargo run | pv > /dev/null`.  This first version
gets about 110 MB/s, but if we wrap it in a BufWriter, it gets 3.5 GB/s!
Thank you to talchass and bluss on Rust IRC for helping with this.

The current version of the code benchmarks @burntsushi's `csv` library,
and returns the following for a nasty CSV with too many columns:

    3000001 rows in 15.95 seconds, 233.63 MiB/sec

CC @seamusabshere