v0.1.3: Map invalid zip codes to empty geochunk ID
Fixes #2. As @wboykinm points out, geodata is commonly messy and broken,
and so we shouldn't abort when encountering bad zip codes. But at the
same time, I don't think it's geochunk's job to clean data or drop bad
records. So instead, we'll simply refrain from assigning a geochunk in
these cases.