Remove obsolete buffering, speed up 5-25%
Now that `csv` supports internal buffering, we no longer need to supply
our own buffers. This simplifies the code, and with a bit of tweaking,
we speed the code up 5 to 25%.

On my laptop, I measure the following speeds:

- Fast path: 125 MiB/sec
- NULL normalization and required columns: 70 MiB/sec
- Above, plus newline removal: 49 MiB/sec

So this remains a tolerably fast program, matching the ~50 MiB/sec that
I expect from many typical C utilities on my laptop.