Optionally remove existing geocode columns
If the geocoder would introduce columns that already exist in the file,
we do one of two things:

1. By default, we exit with an error.
2. If the user passes `--replace`, we strip any column with the same
   name as a geocoding output column, and add the output columns
   normally.

This is fairly tricky code, and I wrote it for a deadline while tired.
So please review it carefully and treat it with caution, and don't use
it as an example of "what to do".