Collapse duplicate suffixes  (fixes #2)
This is a strange pattern that appears in a lot of databases: Sometimes
two address fields will contain redundant data, and sometimes they
won't. I don't really like this code, because it feels hackish, but it
seems to do what people want.

I'm curious about whether we should do something similar with prefixes.