ci: Don't rebuild all binaries for Cargo.lock changes
Technically it makes sense to build all our binaries (but re-release
them) whenever we changed locked dependencies. But that generates a ton
of build spam on CI.

This is arguably reasonable behavior for Rust, which treats Cargo.lock
as not really part of binary crates, and _never_ part of library crates.